+++ extension/src/popup/basics/buttons/BackButton/styles.scss
