+++ extension/src/popup/basics/buttons/PillButton/styles.scss
